added new fastai notebook

+++ advanced/fastai_bring_your_own/container/local_test/deploy_local.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
-image=dogscats-fastai-img
-test_path=$1
+image=$1
+test_path=$2
+
+if [ -z $1 ] || [ -z $2 ]; then
+    echo "ERROR: Must provide a Docker image name and test_path directory"
+    exit 1
+fi
 
 docker run -it -v ${test_path}:/opt/ml -p 8080:8080 --rm ${image} serve
